Adds datalock repo interface Renames updater
