+++ src/SFA.DAS.Commitments.Domain/Data/IDataLockRepository.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using SFA.DAS.Commitments.Domain.Entities.DataLock;
+
+namespace SFA.DAS.Commitments.Domain.Data
+{
+    public interface IDataLockRepository
+    {
+        Task<long> GetLastDataLockEventId();
+        Task<long> UpdateDataLockStatus(DataLockStatus dataLockStatus);
+    }
+}
