+++ src/SFA.DAS.Commitments.Domain/Interfaces/IPaymentEvents.cs
@@ -8,6 +8,6 @@ namespace SFA.DAS.Commitments.Domain.Interfaces
 {
     public interface IPaymentEvents
     {
-        Task<IEnumerable<DataLockStatus>> GetDataLockEvents(int sinceEventId = 0, DateTime? sinceTime = null, string employerAccountId = null, long ukprn = 0, int page = 1);
+        Task<IEnumerable<DataLockStatus>> GetDataLockEvents(long sinceEventId = 0, DateTime? sinceTime = null, string employerAccountId = null, long ukprn = 0, int page = 1);
     }
 }
